requirements.txt
fastapi
uvicorn
python-multipart
sentence-transformers
faiss-cpu
pdfplumber
pymupdf
python-docx
pytesseract
Pillow
pandas
opencv-python
numpy
python-dotenv
openai
streamlit
requests
